adding preliminary post implementation
I have a successful implementation of POST for the database used in the tutorial Jocelyn and I followed. We will need to change variable/database names so it matches Jocelyn's database setup. This syntax worked when I tested it with Postman using http://localhost:8080/warehouses/ as the link.

The other edit I made was removing connection.end() in line 18. This is because when this line was there, I was only able to send 1 request at a time and then the app would need to be restarted.

All the changes in quotes/spacing were auto changes by Prettier.