finishing post request
the post requests are sent in analysis.js using axios. I updated the app.post in index.js to reflect Jocelyn's database.

The remaining problem is that many entries (5-7 usually) are put into the database for each time a single analysis is run. I added the axios post request where Jocelyn had initially attempted to send an SQL query. However, I think this could be the wrong place and maybe we should wait until all data is logged instead of having it in the logdata function.