fixing post request
the axios post request needed to be moved out of logdata and to the end of runAnalysis so that the post request only runs once per analysis (rather than multiple times, since logdata is called multiple times)

I wrote create_sql_data (which formats the analysis data so that it can be sent via post request) based the data structure outline Jocelyn had written in logdata.

I also wrote a separate function send_sql_and_reset, which sends the post request via axios and resets the global variable sql_data (which holds the data that is sent to the db)

After these edits, a single post request will be sent to the db each time you visit a site with the analysis extension.

You must be running the rest api (by node index.js in the terminal) for the post requests to be successful.