updating api and crawler (issue #33)
For the api, you can now do the path /analysis/domain and it will display entries of that domain only. I didn't deploy this update to Cloud Run.

For the cloud crawler, I added
1. some more error catching/timing updates (it seems that if it takes < 13.25 seconds for a site to finish, the driver has turned off)
2. the capability to check whether a site's crawl data has been added to the db (via get request). If it has not been added, the domain gets added to a "try again later list" that will get run at the end of the other sites.

I put it in a separate file since we may still want to use the previous crawler (if we find that the previous failed sites also fail later).