updating crawler and rest api for site_id column (issue #40 )
Jocelyn added site_id as a column in the database. The analysis extension sets the value to null, and then we change it in local-crawler.js with a put request to match the site_id in the csv file.

This was an attempt to write the put requests such that we could run multiple crawlers but just use one table in our analysis database.
While this was not successful (i.e. some site_id's were assigned incorrectly when 2 crawlers were running), it does work correctly when one crawler is running, and with this setup, it would be easy to classify sites as not added or redirect.