adding debugging table and updating local crawler (issue #50)
Since WebDriverError has multiple causes, I separated the reached error page version of WebDriverError so we can isolate it as a site error (rather than a selenium error)

The debugging table is set up such that it is optional. To use it, start the rest api with node index.js debug. Otherwise, start the rest api with node index.js.