integrating privacy pioneer into our crawler (issue #59)
Privacy pioneer is installed after our extension (separately) because it opens a window/tour that we need to click through with selenium.

When an error occurs, we download json data from privacy pioneer before the browser restarts. If the error does not require a restart, we don't download the data. We also download the data at the end after the last site is done.

I got the xpi file by going here: https://addons.mozilla.org/en-US/firefox/addon/privacy-pioneer/ and right clicking on the "Add to Firefox" and using the "Save Link As..." to save it as an xpi file.